Separate project context from environment context
